rrr: convert misc/marshal to module (rrr.marshal)
marshal.cpp becomes the rrr.marshal module interface unit;
marshal.hpp becomes a thin shim that does `import rrr.marshal;`.
Module exports:
  - Marshal class + chunk + bookmark
  - safe_min helper
  - MarshalSink, MarshalSource, MarshalSinkAdapter, MarshalSourceAdapter
    (relocated from serializable in the prior commit)
  - make_sink_proxy(MarshalSink*) / make_source_proxy(MarshalSource*)
  - All `inline operator<<` / `operator>>` Marshal overloads for
    fixed-width ints, v32/v64, strings, pairs, vec/list/set/map (std +
    rusty)

Imports: rrr.basetypes (NoCopy, i8/i16/i32/i64, v32/v64, SparseInt),
rrr.debugging (verify), rrr.serializable (SinkBase, SourceBase,
SinkProxy, SourceProxy), rrr.threading.

Cycle-break: rrr.serializable's prior GMF forward-decls of MarshalSink
/ MarshalSource are now removed — the (removed) convenience
constructors were the only reason for them. rrr.marshal imports
rrr.serializable for the SinkBase/SourceBase interfaces it inherits
from; rrr.serializable does NOT import rrr.marshal. Acyclic.

Consumers updated to import rrr.marshal + rrr.serializable directly
where they previously got them transitively via marshal.hpp:
  - src/rrr/rpc/client.hpp
  - src/rrr/rpc/server.hpp

Fixed: marshal.cpp had `chrono::time_point<...>` from the prior
`using namespace std;` — qualified to `std::chrono::time_point`.

Measurement (rrr + rpcbench, -j32, clean):
  +serializable:  47.42 s, 325 MB PCM   (28 modules)
  +marshal:       61.70 s, 350 MB PCM   (29 modules)

+14 s wallclock — biggest single-step jump in the migration so far,
because marshal is wide fan-out (imported transitively by client.hpp
and through it ~every TU). 3.66× baseline, still under the 5× ceiling.
